+++ raccoontools/shared/file_utils.py
@@ -1,5 +1,5 @@
 from datetime import datetime, UTC
-from typing import Tuple
+from typing import Union
 from uuid import uuid4
 
 
@@ -8,7 +8,7 @@ def get_filename_for_new_file(
         prefix: str = None,
         add_current_datetime_as_format: str = "%Y%m%d%H%M%S%f",
         use_utc: bool = True,
-        unique_identifier: Tuple[str, bool] = True,
+        unique_identifier: Union[str, bool] = True,
         part_separator: str = "-",
         suffix: str = None
 ) -> str: